pyproject.toml
[tool.poetry]
name = "dt_scw_chain"
version = "1.0.0"
description = "Wavetable creation and manipulation tool for Digitakt Single Cycle Waveforms and Chains"
authors = ["Paweł Marczak"] 
repository = "https://github.com/drzhnn/dt_scw_chain.git" 
license = "MIT"
readme = "README.md"
packages = [{ include = "dt_scw_chain", from = "src" }]
classifiers = [
    "Topic :: File Formats",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Sound/Audio :: Conversion",
    "Topic :: Multimedia :: Sound/Audio :: Sound Synthesis"
]


[tool.poetry.dependencies]
python = ">=3.12,<3.14"
numpy = "^1.24.3"
pillow = "^10.4.0"
soundfile = "^0.13.1" 
scipy = "^1.15.0" 

[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
black = "^23.3.0"
ruff = "^0.0.269"
rope = "^1.8.0"
jedi = "^0.18.2"
pylsp-rope = "^0.1.11"
python-lsp-server = {extras = ["all"], version = "^1.7.3"}
ruff-lsp = "^0.0.27"
pylsp-mypy = "^0.6.6"
python-lsp-black = "^1.3.0"
mypy-extensions = "^1.0.0" 

[tool.poetry.scripts]
dt_scw_chain = "dt_scw_chain.app:main"